requirements.txt
accelerate==0.33.0
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
attrs==24.2.0
audioread==3.0.1
bitsandbytes==0.43.3
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==3.0.0
colorama==0.4.6
datasets==2.21.0
decorator==5.1.1
dill==0.3.8
diskcache==5.6.3
distro==1.9.0
docstring_parser==0.16
evaluate==0.4.2
fastapi==0.112.2
filelock==3.15.4
frozenlist==1.4.1
fsspec==2024.6.1
gguf==0.9.1
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.24.6
idna==3.8
importlib_metadata==8.4.0
interegular==0.3.3
Jinja2==3.1.4
jiter==0.5.0
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
lark==1.2.2
lazy_loader==0.4
librosa==0.10.2.post1
llvmlite==0.43.0
lm-format-enforcer==0.10.6
lxml==5.3.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
mpmath==1.3.0
msgpack==1.0.8
msgspec==0.18.6
multidict==6.0.5
multiprocess==0.70.16
nest-asyncio==1.6.0
networkx==3.3
nltk==3.9.1
numba==0.60.0
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-ml-py==12.560.30
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.6.20
nvidia-nvtx-cu12==12.1.105
openai==1.42.0
outlines==0.0.46
packaging==24.1
pandas==2.2.2
peft==0.12.0
pillow==10.4.0
platformdirs==4.2.2
pooch==1.8.2
portalocker==2.10.1
prometheus-fastapi-instrumentator==7.0.0
prometheus_client==0.20.0
protobuf==5.27.3
psutil==6.0.0
py-cpuinfo==9.0.0
pyairports==2.1.1
pyarrow==17.0.0
pycountry==24.6.1
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
Pygments==2.18.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.2
pyzmq==26.2.0
ray==2.34.0
referencing==0.35.1
regex==2024.7.24
requests==2.32.3
rich==13.8.0
rpds-py==0.20.0
sacrebleu==2.4.3
safetensors==0.4.4
scikit-learn==1.5.1
scipy==1.14.1
sentence-transformers==3.0.1
sentencepiece==0.2.0
shtab==1.7.1
six==1.16.0
sniffio==1.3.1
soundfile==0.12.1
soxr==0.5.0
starlette==0.38.2
sympy==1.13.2
tabulate==0.9.0
threadpoolctl==3.5.0
tiktoken==0.7.0
tokenizers==0.19.1
torch==2.4.0
torchvision==0.19.0
tqdm==4.66.5
transformers==4.44.2
triton==3.0.0
trl==0.9.4
typing_extensions==4.12.2
tyro==0.8.10
tzdata==2024.1
urllib3==2.2.2
uvicorn==0.30.6
uvloop==0.20.0
vllm==0.5.5
vllm-flash-attn==2.6.1
watchfiles==0.23.0
websockets==13.0
xformers==0.0.27.post2
xxhash==3.5.0
yarl==1.9.4
zipp==3.20.1
deepspeed==0.15.0
wandb~=0.17.9
together~=1.2.5
anthropic~=0.26.1
torchmetrics~=1.4.0.post0
tree-sitter~=0.22.3
codebleu~=0.7.0
argparse~=1.4.0
aiolimiter~=1.1.0
stable_baselines3~=2.3.2
gymnasium~=0.29.1